+++ frontend/vite.config.js
@@ -3,7 +3,10 @@ import react from '@vitejs/plugin-react'
 
 export default defineConfig({
   plugins: [react()],
+<<<<<<< HEAD
   envDir: '../',
+=======
+>>>>>>> feature/frontend-crud
   server: {
     port: 3000,
   },